+++ Changelog.md
@@ -39,6 +39,7 @@
 - Fix: Floor Trap cloud particles do not spawn when standing at certain positions relative to the Floor Trap
 - Fix: Cloning a passcode-protected block using the /clone command will invalidate the passcode of the original block if the clone is removed
 - Fix: Sonic Security System settings sometimes do not persist through world reloads
+- Fix: The Block Pocket Manager's storage does not persist through world reloads
 - Fix: Potential crash in SaltData (ConcurrentModificationException)
 - Misc.: The minimum required NeoForge version is now 20.4.235